Move display, keyboard and sound to SDL
These functions are related to SDL anyway.
