+++ src/sdl.h
@@ -1,6 +1,6 @@
 /*
  * chip8 is a CHIP-8 emulator done in C
- * Copyright (C) 2015 Dani Rodríguez <danirod@outlook.com>
+ * Copyright (C) 2015-2016 Dani Rodríguez <danirod@outlook.com>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -43,5 +43,29 @@ int init_context(struct context_t*);
 void destroy_context(struct context_t*);
 
 void render(struct context_t*, struct machine_t*);
+
+int is_key_down(char);
+
+/**
+ * Expand a machine_t bitmap into a SDL 32-bit surface bitmap. This function
+ * will work provided some requirements are meet. Both arrays must be 2048
+ * positions long. Input array must be a char array where each position is
+ * either 0 or 1.
+ *
+ * Output array will be expanded in a format that is compatible with SDL
+ * surfaces. Each 0 on input array will be converted to a 0 on output array,
+ * which will result on no pixel being drawn on that position. Each 1 on
+ * input array will be converted to a -1 on output array, which being unsigned
+ * will result on every single bit of that position being 1 and thus
+ * plotting a white pixel.
+ *
+ * @param from input array where the struct machine_t bitmap is from.
+ * @param to output array where the pixels will be blitted.
+ */
+void expand_screen(char*, Uint32*);
+
+SDL_AudioSpec* init_audiospec(void);
+
+void dispose_audiospec(SDL_AudioSpec*);
 
 #endif // SDL_H_
